refactor: extract media store props
