+++ projects/client/src/lib/features/spoilers/useMediaSpoiler.ts
@@ -1,11 +1,13 @@
+import type { MediaStoreProps } from '$lib/models/MediaStoreProps';
 import {
-  type IsWatchedProps,
   useIsWatched,
 } from '$lib/sections/media-actions/mark-as-watched/useIsWatched';
 import { derived } from 'svelte/store';
 import { useSpoiler } from './_internal/useSpoiler';
 
-export function useMediaSpoiler(props: IsWatchedProps) {
+export type MediaSpoilerProps = MediaStoreProps;
+
+export function useMediaSpoiler(props: MediaSpoilerProps) {
   const { isWatched } = useIsWatched(props);
   const { isSpoilerHidden: isHidden } = useSpoiler();
